Corrected several unsigned signal parameters.

+++ include/di/systems/input/input_system.hpp
@@ -190,11 +190,11 @@ public:
   boost::signals2::signal<void(std::string, std::size_t, std::size_t)> on_text_edit              ;
   boost::signals2::signal<void(std::string)>                           on_text_input             ;
   boost::signals2::signal<void()>                                      on_key_layout_change      ;
-  boost::signals2::signal<void(std::array<std::size_t, 2>)>            on_mouse_move             ;
-  boost::signals2::signal<void(std::array<std::size_t, 2>)>            on_mouse_move_delta       ;
+  boost::signals2::signal<void(std::array<std::int32_t, 2>)>           on_mouse_move             ;
+  boost::signals2::signal<void(std::array<std::int32_t, 2>)>           on_mouse_move_delta       ;
   boost::signals2::signal<void(std::size_t)>                           on_mouse_press            ;
   boost::signals2::signal<void(std::size_t)>                           on_mouse_release          ;
-  boost::signals2::signal<void(std::array<std::size_t, 2>)>            on_mouse_wheel            ;
+  boost::signals2::signal<void(std::array<std::int32_t, 2>)>           on_mouse_wheel            ;
   boost::signals2::signal<void(joystick_info)>                         on_joystick_connect       ;
   boost::signals2::signal<void(game_controller_info)>                  on_game_controller_connect;
   boost::signals2::signal<void(std::string)>                           on_clipboard_change       ;
@@ -228,12 +228,12 @@ protected:
         
         else if (event.type == SDL_MOUSEMOTION             )
         {
-          on_mouse_move      ({static_cast<std::size_t>(event.motion.x   ), static_cast<std::size_t>(event.motion.y   )});
-          on_mouse_move_delta({static_cast<std::size_t>(event.motion.xrel), static_cast<std::size_t>(event.motion.yrel)});
+          on_mouse_move      ({static_cast<std::int32_t>(event.motion.x   ), static_cast<std::int32_t>(event.motion.y   )});
+          on_mouse_move_delta({static_cast<std::int32_t>(event.motion.xrel), static_cast<std::int32_t>(event.motion.yrel)});
         }
         else if (event.type == SDL_MOUSEBUTTONDOWN         ) on_mouse_press  ( static_cast<std::size_t>(event.button.button));
         else if (event.type == SDL_MOUSEBUTTONUP           ) on_mouse_release( static_cast<std::size_t>(event.button.button));
-        else if (event.type == SDL_MOUSEWHEEL              ) on_mouse_wheel  ({static_cast<std::size_t>(event.wheel.x), static_cast<std::size_t>(event.wheel.y)});
+        else if (event.type == SDL_MOUSEWHEEL              ) on_mouse_wheel  ({static_cast<std::int32_t>(event.wheel.x), static_cast<std::int32_t>(event.wheel.y)});
         
         else if (event.type == SDL_JOYAXISMOTION           )
         {
@@ -245,7 +245,7 @@ protected:
         {
           auto joystick = std::find_if(joysticks_.begin(), joysticks_.end(), [&event] (const std::unique_ptr<di::joystick>& iteratee) { return iteratee->instance_id() == event.jball.which; });
           if  (joystick == joysticks_.end()) continue;
-          joystick->get()->on_trackball_motion(static_cast<std::size_t>(event.jball.ball), {static_cast<std::size_t>(event.jball.xrel), static_cast<std::size_t>(event.jball.yrel)});
+          joystick->get()->on_trackball_motion(static_cast<std::size_t>(event.jball.ball), {static_cast<std::int32_t>(event.jball.xrel), static_cast<std::int32_t>(event.jball.yrel)});
         }
         else if (event.type == SDL_JOYHATMOTION            ) 
         {
